+++ src/LookupFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Abivia\Geocode;
+
+use Exception;
+
+class LookupFailedException extends Exception
+{
+
+}