+++ src/LookupService/IpApiApi.php
@@ -2,46 +2,50 @@
 
 namespace Abivia\Geocode\LookupService;
 
-use Abivia\Geocode\GeocodeResult\IpInfoResult;
+use Abivia\Geocode\GeocodeResult\IpApiResult;
 use Abivia\Geocode\LookupFailedException;
 
-class IpInfoApi implements LookupService
+class IpApiApi implements LookupService
 {
     /**
      * @var string API access token
      */
-    protected string $token;
+    protected string $accessKey;
 
     /**
-     * @var string Default API base URL (https only on paid plan)
+     * @var string Free base URL
      */
-    protected string $baseUrl = 'https://ipinfo.io/';
+    protected string $baseUrl = 'https://ipapi.co/';
 
     /**
-     * @param string $token
+     * @param string $accessKey
      */
-    public function __construct(string $token = '')
+    public function __construct(string $accessKey = '')
     {
-        $this->token = $token;
+        $this->accessKey = $accessKey;
     }
 
-    public static function make(string $token = ''): static
+    public static function make(string $accessKey = ''): static
     {
-        return new static($token);
+        return new static($accessKey);
     }
 
     /**
-     * Look up the current address via the ioinfo.io API.
+     * Look up the current address via ipapi.co.
      *
      * @param string $address A v4 or v6 IP address.
      * @return array|null
      * @throws LookupFailedException
      */
-    public function query(string $address): ?IpInfoResult
+    public function query(string $address): ?IpApiResult
     {
-        $url = "$this->baseUrl$address/json";
-        if ($this->token !== '') {
-            $url .= '?' . http_build_query(['token' => $this->token]);
+        if ($this->accessKey !== '') {
+            $url = "$this->baseUrl$address/json?"
+                . http_build_query([
+                    'key' => $this->accessKey
+                ]);
+        } else {
+            $url = "$this->baseUrl$address/json";
         }
         $channel = curl_init($url);
         curl_setopt($channel, CURLOPT_RETURNTRANSFER, true);
@@ -53,11 +57,9 @@ class IpInfoApi implements LookupService
                 throw new LookupFailedException("Response was not valid JSON.");
             }
             if ($response['error'] ?? false) {
-                throw new LookupFailedException(
-                    "{$response['error']['title']}: {$response['error']['message']}"
-                );
+                throw new LookupFailedException("{$response['reason']}: {$response['message']}");
             }
-            return new IpInfoResult($response);
+            return new IpApiResult($response);
         }
         return null;
     }