+++ test/GeocodeTest.php
@@ -1,5 +1,6 @@
 <?php
 
+require_once __DIR__ . '/../vendor/autoload.php';
 
 use Abivia\Geocode\AddressNotFoundException;
 use Abivia\Geocode\Geocoder;