+++ test/Live/IpStackTest.php
@@ -17,8 +17,30 @@ class IpStackTest extends TestCase
         $geocoder = new Geocoder(IpStackApi::make(trim(file_get_contents($keyfile))));
         $result = $geocoder->lookup('173.239.198.14');
         $this->assertNotNull($result);
-        $this->assertEquals('US', $result->getCountryCode());
-        $this->assertEquals('173.239.198.14', $result->getIpAddress());
+        $this->assertNotNull($result);
+        $expected = json_decode(
+            file_get_contents(__DIR__ . '/Baseline-173.239.198.14.json'),
+            true
+        );
+        foreach ($expected['exact'] as $method => $value) {
+            if ($method === 'getTimezone') {
+                continue;
+            }
+            $this->assertEquals($value, $result->$method(), $method);
+        }
+        foreach ($expected['close'] as $method => $value) {
+            $this->assertEqualsWithDelta($value, $result->$method(), 0.1, $method);
+        }
+        foreach ($expected['includes'] as $method => $value) {
+            if ($method === 'getLocale') {
+                continue;
+            }
+            $this->assertStringContainsString($value, $result->$method(), $method);
+        }
+        // This provider gets weird postal codes.
+        //foreach ($expected['starts'] as $method => $value) {
+        //    $this->assertStringStartsWith($value, $result->$method(), $method);
+        //}
     }
 
 }